bump to linux-5.10-gen-rkr8, commit-ids:     ./hal/board/.git 280329613bb62f83b064f446fc9c8c205991e111     ./hal/middleware/.git bf7d460318c4539d2a1eef6ec89aaa319e1ad7cb     ./hal/test/.git 0881a7f85277364dc2445f3789ebff52b5617465     ./hal/.git 95edaf1635f8f2a638f94f3dcb52033a05ec08d2     ./hal/applications/.git c5927d3070ebc810b2920f9162bbdada74e63acb     ./uac_app/.git 434c2d15193589fd9960eb910c76333ab3446675     ./rockit/.git 51768b80b2d1f4f9e11f49ee0f38c35e1d38714c     ./drm-cursor/.git 23cebd5dd33a3073f296533954d4c211481891ab     ./libmali/.git 3c3843198ecc0a7a80537fa562faa0b9cfcc3f8a     ./security/bin/.git 5f205632413911a0abe8d13c93bae9dfc72829b7     ./security/rk_tee_user/.git 43cb181927e1357f118c6caeceddaa538da19e06     ./security/librkcrypto/.git 62795797b3aa3b3e9a6c32672a0c4d3d0f44c78a     ./camera_engine_rkaiq/.git 113f1f62c37d6c1153d0d6af70e393438fa43e9a     ./chromium/.git 5ceccf2ddeb9879deb89966ae14ab39c4bec1933     ./rkupdate/.git e9cd4fe07733f807057fd7a6bd5e8d42a59624a7     ./alsa-config/.git 60018d18bdedaccf661c175ab364ae95ebe58a48     ./rockchip-test/.git 4cc483a7b81b3733ecff9b050ea8cebc897abbad     ./xserver/.git b42163ed9f416788301f8bdc5babb219225b2429     ./uvc_app/.git d889984e50904d4615b0fc2668292201e273ebb3     ./recovery/.git dde564ddd12376a8674b3974216cdf37166da5ed     ./rkscript/.git 9c0fc79e9afe73142b2c9557242db7a74be06cac     ./dpdk/.git b9f881cb80adab8e0638b057abe7abf0196515e4     ./rktoolkit/.git bc1d85d6d2a39c67ed2d619f6faf1a996d542722     ./rk_ethercat_release/.git 5fc4657be8f6bf1fc36e4b492cb18885afb56587     ./gstreamer-rockchip/.git 6c90d19b502e36f6228b2d3d31230363bb690b2d     ./rk_pcba_test/.git 9ca36b2bf70f2f164b1891f1ac1ce2bf8eb9325c     ./rkwifibt-app/.git b33a2ca23ca87ccdb64982fac7d0b1791ab714dc     ./common_algorithm/.git 86e6e3e423f28c56692d90cd869798d8e8ce08d0     ./linux-rga/.git 2bff35ced87d66e4412720298f21552f2dc85ae3     ./rkwifibt/.git f4b13f7af66eaa023f942150fb3b1c6e79ddb90e

+++ camera_engine_rkaiq/rkaiq/algos/acac/lut_buffer.h
@@ -36,7 +36,7 @@ struct LutBufferConfig {
     uint32_t Height;
     uint32_t LutHCount;
     uint32_t LutVCount;
-    uint8_t ScaleFactor;
+    uint16_t ScaleFactor;
     uint16_t PsfCfgCount;
 };
